// SDKs/twilio.com/twilio_preview/1.37.3/typescript/src/sdk/models/operations/listsyncdocument.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import * as shared from "../shared";

export const ListSyncDocumentServerList = [
	"https://preview.twilio.com",
] as const;


export class ListSyncDocumentPathParams extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=ServiceSid" })
  serviceSid: string;
}


export class ListSyncDocumentQueryParams extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=PageSize" })
  pageSize?: number;
}


export class ListSyncDocumentSecurity extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "security, scheme=true;type=http;subtype=basic" })
  accountSidAuthToken: shared.SchemeAccountSidAuthToken;
}


export class ListSyncDocumentListSyncDocumentResponseMeta extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=first_page_url" })
  firstPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=key" })
  key?: string;

  @SpeakeasyMetadata({ data: "json, name=next_page_url" })
  nextPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=page" })
  page?: number;

  @SpeakeasyMetadata({ data: "json, name=page_size" })
  pageSize?: number;

  @SpeakeasyMetadata({ data: "json, name=previous_page_url" })
  previousPageUrl?: string;

  @SpeakeasyMetadata({ data: "json, name=url" })
  url?: string;
}


export class ListSyncDocumentListSyncDocumentResponse extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=documents", elemType: shared.PreviewSyncServiceDocument })
  documents?: shared.PreviewSyncServiceDocument[];

  @SpeakeasyMetadata({ data: "json, name=meta" })
  meta?: ListSyncDocumentListSyncDocumentResponseMeta;
}


export class ListSyncDocumentRequest extends SpeakeasyBase {
  @SpeakeasyMetadata()
  serverUrl?: string;

  @SpeakeasyMetadata()
  pathParams: ListSyncDocumentPathParams;

  @SpeakeasyMetadata()
  queryParams: ListSyncDocumentQueryParams;

  @SpeakeasyMetadata()
  security: ListSyncDocumentSecurity;
}


export class ListSyncDocumentResponse extends SpeakeasyBase {
  @SpeakeasyMetadata()
  contentType: string;

  @SpeakeasyMetadata()
  listSyncDocumentResponse?: ListSyncDocumentListSyncDocumentResponse;

  @SpeakeasyMetadata()
  statusCode: number;
}
